requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

a2a-sdk==0.3.25
fastmcp==3.1.1
httpx==0.28.1
langchain==1.2.13
langchain-mcp-adapters==0.2.2
langchain-openai==1.1.12
python-dotenv==1.2.2
requests==2.32.5
uvicorn==0.42.0
